feat: 풀이 글 상세 내 반환 데이터 추가

+++ src/main/java/com/nakaligoba/backend/controller/payload/response/SolutionResponse.java
@@ -23,9 +23,9 @@ public class SolutionResponse {
         private String createdAt;
         private String content;
         private List<String> languages;
-        private int likeCount;
-        private int viewCount;
-        private int commentCount;
+        private Long likeCount;
+        private Long viewCount;
+        private Long commentCount;
         private boolean isLike;
     }
 }
